Fix negative issue in fragment shader within main.py
In the "for i from 0 to 31" loop, I never used "break" when a hit was made. This caused the accumulated color, which was divided by 4 only (for each sample) to always be about 32x what size it should be, resulting in a mostly white hemisphere once clamped to 0 to 1.
Also, antialiasing is working, I think.