Update main.py with antialiasing fixes
Git doesn't like PyCharm's method of shift lines, so the diff looks complex, but only a few lines were changed. The picture now seems to have issues with the specular highlight not reflecting what I would feel should be seen. Such as that there is no specular highlight when the light is between the camera and the sphere. I am not sure of what the cause is.