Add method to load in a pre-computed SDF, pass SDF via 3D texture to shader, and access SDF values in Shader.
Uses moderngl's texture and texture buffer setup (which was difficult to find how to use, oddly enough; have to do "self.texture.use()" before "self.vao.render()" I guess?
Uses sampler3D in the frag shader to get access to the SDF values. I am a bit concerned about how many values are being passed, because I thought it would be 1, but it seemed to show red....
Anyways, a bad picture currently shows instead of the plane, but that shows that the data was passed correctly, hopefully. :smile: